+++ CHANGELOG.md
@@ -1,5 +1,33 @@
 ## [1.2.0] 2018-08-16
 ### Bug fixing
+- Github own repo
+  - [https://github.com/creativetimofficial/material-kit-react/issues/36](https://github.com/creativetimofficial/material-kit-react/issues/36)
+- Github othe repos
+  - [https://github.com/creativetimofficial/ct-material-dashboard-pro-react/issues/79](https://github.com/creativetimofficial/ct-material-dashboard-pro-react/issues/79)
+### Major styling changes
+- `src/assets/jss/material-kit-react/components/headerStyle.jsx`
+- `src/assets/jss/material-kit-react/components/snackbarContentStyle.jsx`
+- `src/assets/jss/material-kit-react/views/loginPage.jsx`
+- `src/assets/jss/material-kit-react/customCheckboxRadioSwitch.jsx`
+- `src/assets/scss/plugins/_plugin-nouislider.scss`
+### Deleted dependencies
+- `react-nouislider`
+### Added dependencies
+- `nouislider`                `12.0.0`
+### Updated dependencies
+- `@material-ui/core`          `1.5.0`   →    `3.1.1`
+- `@material-ui/icons`         `2.0.2`   →    `3.0.1`
+- `react`                     `16.4.2`   →   `16.5.2`
+- `react-dom`                 `16.4.2`   →   `16.5.2`
+- `react-scripts`              `1.1.4`   →    `1.1.5`
+- `react-swipeable-views`    `0.12.16`   →   `0.13.0`
+- `eslint-config-prettier`     `2.9.0`   →    `3.1.0`
+- `eslint-plugin-react`       `7.10.0`   →   `7.11.1`
+- `prettier`                  `1.13.7`   →   `1.14.3`
+
+
+## [1.2.0] 2018-08-16
+### Bug fixing
 - Added new script command for clean install of node_modules (just run in terminal `npm run install:clean`, this will also start your server)
 - Added lint commands
 - Minor changes in components due to the upgrade of `@material-ui/icons`